requirements.txt
django>=5.0.0
django-filter>=23.5
djangorestframework>=3.14.0
djangorestframework>=simplejwt-5.3.1
markdown>=3.5.1
requests>=2.31.0
Whoosh>=2.7.4